generate/requirements.txt
requests~=2.32
transformers~=4.56
torch~=2.8
pillow~=10.4
